refactor: server action to handle cookie and refresh token
